networkd: Stabilise link flap
The e1000e driver flaps the link for ~4-5 seconds during
auto-negotiation on the T14s. This race condition causes
systemd-networkd to tear down the interface immediately after bringing
it up, leading to a stale/broken state.

So as well as the ARP cache fix, where when failing over to Wi-Fi, it
learned the gateway's MAC, and upon replugging Ethernet, the kernel
would incorrectly try to use the Wi-Fi-learned MAC on the wired port, we
need this fix to ride out the hardware link flap.